Moved package command service to the service directory
